Implement UMD commonjsStrict pattern
UMD CommonJS Strict:
  https://github.com/umdjs/umd/blob/master/templates/commonjsStrict.js

This allows the library to define it's module using AMD, CommonJS or a browser
global, usually window.

This changes the pattern slightly to provide the DiceRoller and DiceRoll
directly on the browser global as it was doing before. Therefore there are no
changes to this library when used in a browser however, it can now be used with
AMD or CommonJS as well.