Don't build assets on `npm install`
This causes issues if rollup is not installed, as it's required for asset building, but is a dev dependency.

Instead, we ensure everything is pre-built on `prebuplishOnly`

Also build dev umd file before `npm test`

And use that to run tests on, instead of minified version.
It's quicker to create uncompressed file.